backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
openai==1.51.2
python-dotenv==1.0.1
